+++ script/check-code-style.js
@@ -62,7 +62,8 @@ const components = () => {
         path.basename(dir) !== 'index.ts' &&
         !path.basename(dir).startsWith('.') &&
         !path.basename(dir).includes('signalingprotocol') &&
-        !path.basename(dir).includes('screensignalingprotocol')
+        !path.basename(dir).includes('screensignalingprotocol') &&
+        !path.basename(dir).includes('utils')
     );
 };