+++ test/task/ReleaseMediaStreamsTask.test.ts
@@ -20,13 +20,13 @@ import DefaultStatsCollector from '../../src/statscollector/DefaultStatsCollecto
 import ReleaseMediaStreamsTask from '../../src/task/ReleaseMediaStreamsTask';
 import Task from '../../src/task/Task';
 import DefaultTransceiverController from '../../src/transceivercontroller/DefaultTransceiverController';
+import { wait as delay } from '../../src/utils/Utils';
 import NoVideoDownlinkBandwidthPolicy from '../../src/videodownlinkbandwidthpolicy/NoVideoDownlinkBandwidthPolicy';
 import DefaultVideoTileController from '../../src/videotilecontroller/DefaultVideoTileController';
 import DefaultVideoTileFactory from '../../src/videotilefactory/DefaultVideoTileFactory';
 import DefaultWebSocketAdapter from '../../src/websocketadapter/DefaultWebSocketAdapter';
 import DOMMockBehavior from '../dommock/DOMMockBehavior';
 import DOMMockBuilder from '../dommock/DOMMockBuilder';
-import { delay } from '../utils';
 
 describe('ReleaseMediaStreamsTask', () => {
   const expect: Chai.ExpectStatic = chai.expect;